@@ -6,21 +6,39 @@ import android.app.Dialog;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.ads.nativetemplates.NativeTemplateStyle;
+import com.google.android.ads.nativetemplates.TemplateView;
+import com.google.android.gms.ads.AdLoader;
+import com.google.android.gms.ads.AdRequest;
+
 import java.util.Random;
 
+import go.app.newe.App;
 import go.app.newe.R;
+import go.app.newe.data.a.model.Advertisement;
+import go.app.newe.data.a.model.Screen;
+import go.app.newe.data.a.model.ViewItem;
 import go.app.newe.list.Data_Buttons;
 
 public class Art_5 extends AppCompatActivity {
 
     private Dialog dialog;
     private Button ShowDialog;
+
+    LinearLayout nativeContainer;
+
+    Handler handler = new Handler();
+    private Screen mScreen;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -37,9 +55,7 @@ public class Art_5 extends AppCompatActivity {
         //Create the Dialog here
         dialog = new Dialog(this);
         dialog.setContentView(R.layout.custom_dialog_layout);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            dialog.getWindow().setBackgroundDrawable(getDrawable(R.drawable.custom_dialog_background));
-        }
+        dialog.getWindow().setBackgroundDrawable(getDrawable(R.drawable.custom_dialog_background));
         dialog.getWindow().setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT);
         dialog.setCancelable(false); //Optional
         dialog.getWindow().getAttributes().windowAnimations = R.style.DialogAnimation; //Setting the animations to dialog
@@ -73,6 +89,61 @@ public class Art_5 extends AppCompatActivity {
                 dialog.show(); // Showing the dialog here
             }
         });
+
+
+        nativeContainer = findViewById(R.id.native_container);
+        setupView();
+    }
+
+    private void setupView() {
+        for (Screen screen : App.getAppConfig().getScreens()) {
+            if (screen.getName().equals("article_5_screen")) {
+                mScreen = screen;
+                break;
+            }
+        }
+
+        for (ViewItem item : mScreen.getViewItems()) {
+            if (item.getName().equals("body_1") && item.getType().equals("textView")) {
+                TextView body = findViewById(R.id.body_1);
+                body.setText(item.getText());
+            } else if (item.getName().equals("body_2") && item.getType().equals("textView")) {
+                TextView body = findViewById(R.id.body_2);
+                body.setText(item.getText());
+            } else if (item.getName().equals("title") && item.getType().equals("textView")) {
+                TextView title = findViewById(R.id.title);
+                title.setText(item.getText());
+            }
+        }
+
+        handler.post(() -> {
+            if (mScreen != null) {
+                for (Advertisement advertisement : mScreen.getAdvertisements()) {
+                    if (advertisement.getProvider().equals("admob")
+                            && advertisement.getType().equals("native")
+                            && advertisement.getName().equals("native_1")
+                            && advertisement.getEnabled()
+                            && advertisement.getAdId() != null) {
+                        Log.d("TAG", "setupView: ad id => " + advertisement.getAdId());
+                        AdLoader adLoader = new AdLoader.Builder(Art_5.this, advertisement.getAdId())
+                                .forNativeAd(nativeAd -> {
+                                    NativeTemplateStyle styles = new NativeTemplateStyle
+                                            .Builder()
+                                            .build();
+                                    LayoutInflater layoutInflater = (LayoutInflater) Art_5.this.getSystemService(LAYOUT_INFLATER_SERVICE);
+                                    View view = layoutInflater.inflate(R.layout.medium_template_view, nativeContainer);
+                                    TemplateView template = view.findViewById(R.id.my_template);
+                                    template.setStyles(styles);
+                                    template.setNativeAd(nativeAd);
+                                })
+                                .build();
+                        adLoader.loadAd(new AdRequest.Builder().build());
+
+                    }
+                }
+            }
+        });
+
     }
 
     public void back(View view) {
