@@ -1,13 +1,18 @@
 package go.app.newe.pages;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.os.Handler;
+import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
@@ -17,13 +22,29 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
+import com.google.android.ads.nativetemplates.NativeTemplateStyle;
+import com.google.android.ads.nativetemplates.TemplateView;
+import com.google.android.gms.ads.AdError;
+import com.google.android.gms.ads.AdLoader;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.FullScreenContentCallback;
+import com.google.android.gms.ads.LoadAdError;
+import com.google.android.gms.ads.interstitial.InterstitialAd;
+import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
 
 import br.com.simplepass.loading_button_lib.customViews.CircularProgressButton;
+import go.app.newe.App;
 import go.app.newe.R;
+import go.app.newe.data.a.model.Advertisement;
+import go.app.newe.data.a.model.Screen;
+import go.app.newe.data.a.model.ViewItem;
+import go.app.newe.list.DialogN;
 
 public class Activity_5 extends AppCompatActivity {
 
@@ -36,8 +57,16 @@ public class Activity_5 extends AppCompatActivity {
     ImageView photoIV;
     CircularProgressButton circularProgressButton;
 
+    private DialogN dialogN = DialogN.getInstance();
+
     LinearLayout bannerContainer;
     LinearLayout nativeContainer;
+
+
+    private InterstitialAd mInterstitialAd;
+
+    Handler handler = new Handler();
+    private Screen mScreen;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,7 +83,7 @@ public class Activity_5 extends AppCompatActivity {
         circularProgressButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                AsyncTask<String, String, String> demoDownload = new AsyncTask<String, String, String>() {
+                @SuppressLint("StaticFieldLeak") AsyncTask<String, String, String> demoDownload = new AsyncTask<String, String, String>() {
                     @Override
                     protected String doInBackground(String... strings) {
                         try {
@@ -68,13 +97,8 @@ public class Activity_5 extends AppCompatActivity {
                     @Override
                     protected void onPostExecute(String s) {
                         if (s.equals("done")) {
-                            Intent intent = new Intent(Activity_5.this, Activity_6.class);
-                            startActivity(intent);
-
-
+                            showInterstitial();
                             circularProgressButton.doneLoadingAnimation(Color.parseColor("#333639"), BitmapFactory.decodeResource(getResources(), R.drawable.ic_done_white_48dp));
-
-
                         }
 
                     }
@@ -85,26 +109,131 @@ public class Activity_5 extends AppCompatActivity {
         });
 
         photoIV = findViewById(R.id.photo);
+        photoIV.setOnClickListener((v) -> {
+            if (dialogN == null) {
+                dialogN = DialogN.getInstance();
+            }
+            dialogN.setListener(() -> {
+                if (dialogN != null)
+                    dialogN.dismiss();
+                Glide.with(this).load(App.getDataManager().getUserImage()).into(photoIV);
+            });
+            dialogN.show(getSupportFragmentManager(), "A_TAG");
+        });
 
-        // Configure sign-in to request the user's ID, email address, and basic
-        // profile. ID and basic profile are included in DEFAULT_SIGN_IN.
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestEmail()
-                .build();
-
-        // Build a GoogleSignInClient with the options specified by gso.
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
-
-        GoogleSignInAccount acct = GoogleSignIn.getLastSignedInAccount(Activity_5.this);
-        if (acct != null) {
-
-            Uri personPhoto = acct.getPhotoUrl();
-
-            Glide.with(this).load(personPhoto).into(photoIV);
-        }
+        Glide.with(this).load(App.getDataManager().getUserImage()).into(photoIV);
     }
 
     private void setupAds() {
-        // TODO ====>
+        for (Screen screen : App.getAppConfig().getScreens()) {
+            if (screen.getName().equals("step_4_screen")) {
+                mScreen = screen;
+                break;
+            }
+        }
+
+        for (ViewItem item : mScreen.getViewItems()) {
+            if (item.getName().equals("body_1") && item.getType().equals("textView")) {
+                TextView body = findViewById(R.id.body);
+                body.setText(item.getText());
+            }
+        }
+
+
+        handler.post(() -> {
+            if (mScreen != null) {
+                for (Advertisement advertisement : mScreen.getAdvertisements()) {
+                    if (advertisement.getProvider().equals("admob")
+                            && advertisement.getType().equals("native")
+                            && advertisement.getName().equals("native_1")
+                            && advertisement.getEnabled()
+                            && advertisement.getAdId() != null) {
+                        Log.d("TAG", "setupView: ad id => " + advertisement.getAdId());
+                        AdLoader adLoader = new AdLoader.Builder(Activity_5.this, advertisement.getAdId())
+                                .forNativeAd(nativeAd -> {
+                                    NativeTemplateStyle styles = new NativeTemplateStyle
+                                            .Builder()
+                                            .build();
+                                    LayoutInflater layoutInflater = (LayoutInflater) Activity_5.this.getSystemService(LAYOUT_INFLATER_SERVICE);
+                                    View view = layoutInflater.inflate(R.layout.medium_template_view, nativeContainer);
+                                    TemplateView template = view.findViewById(R.id.my_template);
+                                    template.setStyles(styles);
+                                    template.setNativeAd(nativeAd);
+                                })
+                                .build();
+                        adLoader.loadAd(new AdRequest.Builder().build());
+
+                    }
+
+                    if (advertisement.getProvider().equals("admob")
+                            && advertisement.getType().equals("interstitial")
+                            && advertisement.getName().equals("interstitial_1")
+                            && advertisement.getEnabled()
+                            && advertisement.getAdId() != null) {
+                        AdRequest adRequest = new AdRequest.Builder().build();
+                        InterstitialAd.load(Activity_5.this, advertisement.getAdId(), adRequest,
+                                new InterstitialAdLoadCallback() {
+                                    @Override
+                                    public void onAdLoaded(@NonNull InterstitialAd interstitialAd) {
+                                        mInterstitialAd = interstitialAd;
+                                        setListener();
+                                    }
+
+                                    @Override
+                                    public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
+                                        // Handle the error
+                                        Log.i("TAG", loadAdError.getMessage());
+                                        mInterstitialAd = null;
+                                    }
+                                });
+                    }
+
+                    if (advertisement.getProvider().equals("admob")
+                            && advertisement.getType().equals("banner")
+                            && advertisement.getName().equals("banner_1")
+                            && advertisement.getEnabled()
+                            && advertisement.getAdId() != null) {
+                        AdView adView = new AdView(this);
+                        adView.setAdSize(AdSize.SMART_BANNER);
+                        adView.setAdUnitId(advertisement.getAdId());
+                        bannerContainer.addView(adView);
+                        adView.loadAd(new AdRequest.Builder().build());
+                    }
+                }
+            }
+        });
+
+
+    }
+
+    private void setListener() {
+        if (mInterstitialAd != null) {
+            mInterstitialAd.setFullScreenContentCallback(new FullScreenContentCallback() {
+                @Override
+                public void onAdDismissedFullScreenContent() {
+                    startActivity(new Intent(Activity_5.this, Activity_6.class));
+                    finish();
+                }
+
+                @Override
+                public void onAdFailedToShowFullScreenContent(@NonNull AdError adError) {
+                }
+
+                @Override
+                public void onAdShowedFullScreenContent() {
+                    mInterstitialAd = null;
+                    Log.d("TAG", "The ad was shown.");
+                }
+            });
+        }
+    }
+
+    private void showInterstitial() {
+        if (mInterstitialAd != null) {
+            mInterstitialAd.show(this);
+        } else {
+            startActivity(new Intent(Activity_5.this, Activity_6.class));
+            finish();
+        }
     }
 }
